+++ examples/classic_controllers/classic_controllers_synch_motor_example.py
@@ -50,14 +50,13 @@ if __name__ == '__main__':
 
     controller = Controller.make(env, external_ref_plots=external_ref_plots, torque_control='analytical')
 
-    state, reference = env.reset()
+    (state, reference), _ = env.reset()
 
     # simulate the environment
     for i in range(10001):
-        env.render()
         action = controller.control(state, reference)
-        (state, reference), reward, done, _ = env.step(action)
-        if done:
+        (state, reference), reward, terminated, truncated, _ = env.step(action)
+        if terminated:
             env.reset()
             controller.reset()