+++ examples/classic_controllers/custom_classic_controllers_dc_motor_example.py
@@ -39,7 +39,7 @@ if __name__ == '__main__':
     external_ref_plots = [ExternallyReferencedStatePlot(state) for state in states]
 
     # initialize the gym-electric-motor environment
-    env = gem.make(motor, visualization=MotorDashboard(additional_plots=external_ref_plots))
+    env = gem.make(motor, visualization=MotorDashboard(additional_plots=external_ref_plots), render_mode = 'figure')
 
     """
         initialize the controller
@@ -69,14 +69,13 @@ if __name__ == '__main__':
 
     controller = Controller.make(env, external_ref_plots=external_ref_plots, stages=stages)
 
-    state, reference = env.reset()
+    (state, reference), _ = env.reset()
 
     # simulate the environment
     for i in range(10001):
         action = controller.control(state, reference)
-        env.render()
-        (state, reference), reward, done, _ = env.step(action)
-        if done:
+        (state, reference), reward, terminated, truncated, _ = env.step(action)
+        if terminated:
             env.reset()
             controller.reset()