+++ examples/classic_controllers/custom_classic_controllers_synch_motor_example.py
@@ -35,7 +35,8 @@ if __name__ == '__main__':
 
     # initialize the gym-electric-motor environment
     env = gem.make(env_id, visualization=MotorDashboard(additional_plots=external_ref_plots),
-                   motor=dict(limit_values=limit_values, nominal_values=nominal_values, motor_parameter=motor_parameter))
+                   motor=dict(limit_values=limit_values, nominal_values=nominal_values, motor_parameter=motor_parameter),
+                   render_mode = 'figure')
 
     """
     initialize the controller
@@ -100,14 +101,13 @@ if __name__ == '__main__':
 
     controller = Controller.make(env, stages=stages, external_ref_plots=external_ref_plots, torque_control='analytical')
 
-    state, reference = env.reset()
+    (state, reference), _ = env.reset()
 
     # simulate the environment
     for i in range(10001):
-        env.render()
         action = controller.control(state, reference)
-        (state, reference), reward, done, _ = env.step(action)
-        if done:
+        (state, reference), reward, terminated, truncated, _ = env.step(action)
+        if terminated:
             env.reset()
             controller.reset()