+++ gym_electric_motor/envs/gym_dcm/extex_dc_motor_env/cont_cc_extex_dc_env.py
@@ -76,16 +76,15 @@ class ContCurrentControlDcExternallyExcitedMotorEnv(ElectricMotorEnvironment):
         ...     ode_solver=my_overridden_solver,
         ...     reference_generator=my_new_ref_gen_instance
         ... )
-        >>> done = True
+        >>> terminated = True
         >>> for _ in range(1000):
-        >>>     if done:
+        >>>     if terminated:
         >>>         state, reference = env.reset()
-        >>>     env.render()
-        >>>     (state, reference), reward, done, _ = env.step(env.action_space.sample())
+        >>>     (state, reference), reward, terminated, truncated, _ = env.step(env.action_space.sample())
     """
     def __init__(self, supply=None, converter=None, motor=None, load=None, ode_solver=None,
                  reward_function=None, reference_generator=None, visualization=None, state_filter=None, callbacks=(),
-                 constraints=('i_a', 'i_e'), calc_jacobian=True, tau=1e-4, physical_system_wrappers=()):
+                 constraints=('i_a', 'i_e'), calc_jacobian=True, tau=1e-4, physical_system_wrappers=(), **kwargs):
         """
         Args:
             supply(env-arg): Specification of the :py:class:`.VoltageSupply` for the environment
@@ -156,5 +155,5 @@ class ContCurrentControlDcExternallyExcitedMotorEnv(ElectricMotorEnvironment):
         super().__init__(
             physical_system=physical_system, reference_generator=reference_generator, reward_function=reward_function,
             constraints=constraints, visualization=visualization, state_filter=state_filter, callbacks=callbacks,
-            physical_system_wrappers=physical_system_wrappers
+            physical_system_wrappers=physical_system_wrappers, **kwargs
         )