+++ gym_electric_motor/envs/gym_eesm/finite_cc_eesm_env.py
@@ -77,16 +77,16 @@ class FiniteCurrentControlExternallyExcitedSynchronousMotorEnv(ElectricMotorEnvi
         ...     ode_solver=my_overridden_solver,
         ...     reference_generator=my_new_ref_gen_instance
         ... )
-        >>> done = True
+        >>> terminated = True
         >>> for _ in range(1000):
-        >>>     if done:
+        >>>     if terminated:
         >>>         state, reference = env.reset()
-        >>>     env.render()
-        >>>     (state, reference), reward, done, _ = env.step(env.action_space.sample())
+        >>>     (state, reference), reward, terminated, truncated, _ = env.step(env.action_space.sample())
     """
     def __init__(self, supply=None, converter=None, motor=None, load=None, ode_solver=None,
                  reward_function=None, reference_generator=None, visualization=None, state_filter=None, callbacks=(),
-                 constraints=(SquaredConstraint(('i_sq', 'i_sd')), LimitConstraint(('i_e',))), calc_jacobian=True, tau=1e-5):
+                 constraints=(SquaredConstraint(('i_sq', 'i_sd')), LimitConstraint(('i_e',))), calc_jacobian=True,
+                 tau=1e-5, physical_system_wrappers=(), **kwargs):
         """
         Args:
             supply(env-arg): Specification of the :py:class:`.VoltageSupply` for the environment
@@ -107,6 +107,8 @@ class FiniteCurrentControlExternallyExcitedSynchronousMotorEnv(ElectricMotorEnvi
             tau(float): Duration of one control step in seconds. Default: 1e-4.
             state_filter(list(str)): List of states that shall be returned to the agent. Default: None (no filter)
             callbacks(list(Callback)): Callbacks for user interaction. Default: ()
+            physical_system_wrappers(list(PhysicalSystemWrapper)): List of Physical System Wrappers to modify the
+            actions to and states from the physical system before they are used in the environment. Default: ()
 
         Note on the env-arg type:
             All parameters of type env-arg can be selected as one of the following types:
@@ -156,5 +158,6 @@ class FiniteCurrentControlExternallyExcitedSynchronousMotorEnv(ElectricMotorEnvi
         )
         super().__init__(
             physical_system=physical_system, reference_generator=reference_generator, reward_function=reward_function,
-            constraints=constraints, visualization=visualization, state_filter=state_filter, callbacks=callbacks
+            constraints=constraints, visualization=visualization, state_filter=state_filter, callbacks=callbacks,
+            physical_system_wrappers=physical_system_wrappers, **kwargs
         )