+++ gym_electric_motor/physical_systems/physical_systems.py
@@ -1,5 +1,5 @@
 import numpy as np
-from gym.spaces import Box
+from gymnasium.spaces import Box
 import warnings
 
 import gym_electric_motor as gem