+++ tests/testing_utils.py
@@ -8,7 +8,7 @@ from gym_electric_motor.physical_systems import PowerElectronicConverter, Mechan
 import gym_electric_motor.physical_systems.converters as cv
 from gym_electric_motor.physical_systems.physical_systems import SCMLSystem
 import numpy as np
-from gym.spaces import Box, Discrete
+from gymnasium.spaces import Box, Discrete
 from scipy.integrate import ode
 from tests.conf import system, jacobian, permex_motor_parameter
 from gym_electric_motor.utils import instantiate
@@ -199,7 +199,7 @@ class DummyRewardFunction(RewardFunction):
         self.last_action = None
         self.last_time_step = None
         self.closed = False
-        self.done = False
+        self.terminated = False
         self.kwargs = kwargs
         super().__init__()