+++ CHANGELOG.md
@@ -7,11 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ## Added
 - Support for Python 3.10
+- Changed API to be compatible with Gymnasium v0.29.0
+- Output a Motor Dashborad graph as PNG or PDF file (with LaTeX font)
+- Automated integration test (run a motor simulation with controller)
+
+
 
 ## Changed
 - Dropped support for Python 3.7 (due to gymnasium not supporting it)
 
 ## Fixed
+- Bugfixes for EESM environment
 
 ## [1.1.0] - 2022-04-25
 ## Added